Map square input to circle
I knew there was a simple way to do it. We just needed to make the total speed the higher of the 2 speed components. Changing joystick position along 1 axis while staying on one edge of the same square (by keeping the other the same and not having the moving one exceed the still one's value) will keep the speed the same while changing the angle, as if you were moving along the edge of a circle. It should turn squares into concentric circles like expected, and it works intuitively but I also have equations that prove it a little more rigorously.